Move `get_current_release_version` from NewVersionAvailable.py to web.py
