+++ src/sas/qtgui/Utilities/NewVersion/NewVersionAvailable.py
@@ -2,8 +2,7 @@ import logging
 import webbrowser
 from copy import copy
 
-import requests
-from packaging.version import Version, parse
+from packaging.version import parse
 from PySide6.QtCore import QSize
 from PySide6.QtGui import QIcon
 from PySide6.QtWidgets import (
@@ -19,8 +18,8 @@ from PySide6.QtWidgets import (
 )
 
 from sas import config
-from sas.system import web
 from sas.system.version import __version__ as current_version_string
+from sas.system.web import get_current_release_version
 
 logger = logging.getLogger("NewVersionAvailable")
 
@@ -81,27 +80,6 @@ class NewVersionAvailable(QDialog):
         if not self.dont_show.isChecked():
             config.LAST_UPDATE_DISMISSED_VERSION = self.latest_version
         self.close()
-
-
-def get_current_release_version() -> tuple[str, str, Version] | None:
-    """ Get the current version from the server """
-    try:
-        response = requests.get(web.update_url, timeout=config.UPDATE_TIMEOUT)
-        # Will throw Exception if the HTTP status code returned isn't success
-        # (2xx)
-        response.raise_for_status()
-        version_info = response.json()
-        logger.info("Connected to www.sasview.org. Received: %s", version_info)
-
-        version_string = version_info["version"]
-        url = version_info["download_url"]
-
-        return version_string, url, parse(version_string)
-
-    except Exception as ex:
-        logger.info("Failed to get version number %s", ex)
-        return None
-
 
 
 def maybe_prompt_new_version_download() -> QDialog | None:
